# Add memtable per key-value checksum