--- a/tools/db_crashtest.py
+++ b/tools/db_crashtest.py
@@ -35,6 +35,8 @@ default_params = {
     # Consider larger number when backups considered more stable
     "backup_one_in": 100000,
     "batch_protection_bytes_per_key": lambda: random.choice([0, 8]),
+    # 1 is too flaky
+    "memtable_protection_bytes_per_key": lambda: random.choice([0, 2, 4, 8]),
     "block_size": 16384,
     "bloom_bits": lambda: random.choice([random.randint(0,19),
                                          random.lognormvariate(2.3, 1.3)]),