… HISTORY.md
… db_stress_tool/db_stress_test_base.cc
… options/options_settable_test.cc
… tools/db_bench_tool.cc